Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/webaudio/the-audio-api/the-delaynode-interface/no-dezippering-expected.txt

CONSOLE MESSAGE: line 24: SyntaxError: The string did not match the expected pattern.

Harness Error (FAIL), message = SyntaxError: The string did not match the expected pattern.

PASS # AUDIT TASK RUNNER STARTED. 
PASS > [test0] Test DelayNode has no dezippering 
